Navbar has been made dynamic
