+++ src/pages/buyer/cart/style.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/variables.scss";
+@import "../../../assets/styles/variables.scss";
 
 .shopping-cart-page {
   h2 {
