+++ src/pages/buyer/checkout/style.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/variables.scss";
+@import "../../../assets/styles/variables.scss";
 
 .checkout-form-container {
   max-width: 50%;
