+++ src/pages/buyer/home/index.js
@@ -1,7 +1,7 @@
-import NavScrollExample from "../../layout/navbar/index";
-import HeroBanner from "../../components/banner/index";
-import CateCard from "../../components/cateCard/index";
-import FeaturedCard from "../../components/featuredCard";
+import NavScrollExample from "../../../layout/navbar/index";
+import HeroBanner from "../../../components/banner/index";
+import CateCard from "../../../components/cateCard/index";
+import FeaturedCard from "../../../components/featuredCard";
 import productData from "../constants/productData";
 import { FaHeart, FaCakeCandles, FaRing, FaTree } from "react-icons/fa6";
