+++ src/pages/buyer/itemDescription/style.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/variables.scss";
+@import "../../../assets/styles/variables.scss";
 
 .product-description {
   max-width: 1200px;
