+++ src/pages/buyer/review/index.js
@@ -1,9 +1,9 @@
 import { useLocation } from "react-router-dom";
 import { useState, useEffect } from "react";
-import NavScrollExample from "../../layout/navbar";
-import CustomButton from "../../components/button/index";
+import NavScrollExample from "../../../layout/navbar";
+import CustomButton from "../../../components/button/index";
 import Form from "react-bootstrap/Form";
-import ReviewCard from "../../components/reviewCard";
+import ReviewCard from "../../../components/reviewCard";
 import "./style.scss";
 
 const ReviewPage = ({ title = "Customer Reviews" }) => {
