+++ src/pages/buyer/review/style.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/variables.scss";
+@import "../../../assets/styles/variables.scss";
 
 .review-form-container {
   max-width: 50%;
